codebase(server): reformat server origanize
