+++ config/environments/demo.yaml
@@ -5,7 +5,7 @@ models:
     temperature: 0.7
     max_tokens: 1000
     timeout: 30
-  
+
   model2:
     name: "gpt-4o"
     type: "openai"
@@ -19,14 +19,14 @@ models:
   #   type: anthropic
   #   temperature: 0.7
   #   timeout: 45
-  
+
   # Together AI models (uncomment to use)
   # llama-2-70b-together:
   #   name: meta-llama/Llama-2-70b-chat-hf
   #   type: together
   #   temperature: 0.7
   #   timeout: 45
-  
+
   # vLLM models (uncomment to use)
   # llama-2-7b-vllm:
   #   name: llama-2-7b-chat
@@ -41,7 +41,7 @@ evaluators:
       max_turn_number: 20
       max_stale_turn: 2
     enabled: true
-  
+
   - type: "llm"
     model: "model2"
     config:
