+++ config/environments/workflow_provider_config.yaml
@@ -1,4 +1,3 @@
-
 models:
   simple-custom:
     name: simple-custom
