+++ tiny_chat/__init__.py
@@ -1,30 +1,14 @@
 __version__ = '0.0.1'
 
 from .agents import BaseAgent, HumanAgent, LLMAgent
+from .config import ConfigManager, ModelProviderConfig
 from .envs import TinyChatEnvironment
 from .evaluator import (
-    EpisodeLLMEvaluator,
+    BaseEvaluator,
     EvaluationForMultipleAgents,
-    Evaluator,
-    RuleBasedTerminatedEvaluator,
+    LLMEvaluator,
+    RuleBasedEvaluator,
     unweighted_aggregate_evaluate,
-)
-from .generator import (
-    EnvResponse,
-    ListOfIntOutputParser,
-    PydanticOutputParser,
-    ScriptOutputParser,
-    StrOutputParser,
-    agenerate,
-    agenerate_action,
-    agenerate_env_profile,
-    agenerate_goal,
-    agenerate_init_profile,
-    agenerate_relationship_profile,
-    agenerate_script,
-    convert_narratives,
-    format_bad_output,
-    process_history,
 )
 from .messages import (
     ActionType,
@@ -46,22 +30,35 @@ from .profiles import (
     FineGrainedRelationshipProfile,
     RelationshipType,
 )
-from .server import (
-    ConfigManager,
-    ModelProviderConfig,
+from .providers import (
+    EnvResponse,
+    ListOfIntOutputParser,
     ModelProviderFactory,
-    TinyChatServer,
+    PydanticOutputParser,
+    ScriptOutputParser,
+    StrOutputParser,
+    agenerate,
+    agenerate_action,
+    agenerate_env_profile,
+    agenerate_goal,
+    agenerate_init_profile,
+    agenerate_relationship_profile,
+    agenerate_script,
+    convert_narratives,
+    format_bad_output,
+    process_history,
 )
+from .server import TinyChatServer
 
 __all__ = [
     'LLMAgent',
     'BaseAgent',
     'HumanAgent',
     'TinyChatEnvironment',
-    'EpisodeLLMEvaluator',
+    'BaseEvaluator',
     'EvaluationForMultipleAgents',
-    'Evaluator',
-    'RuleBasedTerminatedEvaluator',
+    'LLMEvaluator',
+    'RuleBasedEvaluator',
     'unweighted_aggregate_evaluate',
     'ActionType',
     'AgentAction',
