+++ tiny_chat/envs/__init__.py
@@ -1,3 +1,4 @@
-from .environment import TinyChatEnvironment
+from .base import BaseChatEnivronment
+from .tinychat_env import TinyChatEnvironment
 
-__all__ = ['TinyChatEnvironment']
+__all__ = ['TinyChatEnvironment', 'BaseChatEnivronment']
