+++ tiny_chat/envs/tinychat_env.py
@@ -1,12 +1,11 @@
 import asyncio
 import itertools
 import random
-from abc import ABC, abstractmethod
 from typing import Any, Literal, TypeVar
 
 from pydantic import validate_call
 
-from tiny_chat.evaluator import Evaluator, unweighted_aggregate_evaluate
+from tiny_chat.evaluator import BaseEvaluator, unweighted_aggregate_evaluate
 from tiny_chat.messages import (
     ActionType,
     AgentAction,
@@ -16,6 +15,8 @@ from tiny_chat.messages import (
     SimpleMessage,
     TinyChatBackground,
 )
+
+from .base import BaseChatEnivronment
 
 TBackground = TypeVar('TBackground', bound=ScriptBackground)
 
@@ -38,36 +39,6 @@ def _actions_to_natural_language(actions: dict[str, AgentAction]) -> str:
     return action_str
 
 
-class BaseChatEnivronment(ABC):
-    @abstractmethod
-    def reset(self, **kwargs: Any) -> Any:
-        pass
-
-    @abstractmethod
-    def step(
-        self, actions: dict[str, AgentAction] | dict[str, dict[str, int | str]]
-    ) -> Any:
-        pass
-
-    @abstractmethod
-    def astep(
-        self, actions: dict[str, AgentAction] | dict[str, dict[str, int | str]]
-    ) -> Any:
-        pass
-
-    @abstractmethod
-    def get_observation(self, agent_name: str) -> Observation:
-        pass
-
-    @abstractmethod
-    def is_terminated(self) -> bool:
-        pass
-
-    @abstractmethod
-    def get_turn_number(self) -> int:
-        pass
-
-
 class TinyChatEnvironment(BaseChatEnivronment):
     def __init__(
         self,
@@ -75,9 +46,9 @@ class TinyChatEnvironment(BaseChatEnivronment):
         action_order: Literal[
             'simultaneous', 'round-robin', 'sequential', 'random', 'agent_id_based'
         ] = 'simultaneous',
-        evaluators: list[Evaluator] | None = None,
+        evaluators: list[BaseEvaluator] | None = None,
         model_name: str = 'gpt-4o-mini',
-        terminal_evaluators: list[Evaluator] | None = None,
+        terminal_evaluators: list[BaseEvaluator] | None = None,
         max_turns: int = 20,
         speaking_order: list[int] | None = None,
         obs_mode: Literal['all', 'local'] = 'all',
@@ -461,16 +432,21 @@ class TinyChatEnvironment(BaseChatEnivronment):
         observations = {}
         for i, agent_name in enumerate(self.agent_names):
             observations[agent_name] = Observation(
-                last_turn=self._last_turn_text_for(agent_name)
-                if self.turn_number > 0
-                else (
-                    (self.env_background or self.background)
-                    .create_agent_specific_background(  # type: ignore
-                        agent_name, omniscient=False
+                last_turn=(
+                    self._last_turn_text_for(agent_name)
+                    if self.turn_number > 0
+                    else (
+                        (self.env_background or self.background)
+                        .create_agent_specific_background(  # type: ignore
+                            agent_name, omniscient=False
+                        )
+                        .to_natural_language()
+                        if not self._omniscient
+                        and (self.env_background or self.background)
+                        else (
+                            self.env_background or self.background
+                        ).to_natural_language()  # type: ignore
                     )
-                    .to_natural_language()
-                    if not self._omniscient and (self.env_background or self.background)
-                    else (self.env_background or self.background).to_natural_language()  # type: ignore
                 ),
                 turn_number=self.turn_number,
                 available_actions=(
