+++ tiny_chat/evaluator/__init__.py
@@ -1,18 +1,17 @@
+from .base import BaseEvaluator
 from .dimensions import SotopiaDimensions
-from .evaluators import (
-    EpisodeLLMEvaluator,
-    EvaluationForMultipleAgents,
-    Evaluator,
-    RuleBasedTerminatedEvaluator,
-    unweighted_aggregate_evaluate,
-)
+from .llm_evaluator import LLMEvaluator
+from .manager import EvaluatorConfig, EvaluatorManager
+from .rule_based import RuleBasedEvaluator
+from .utils import EvaluationForMultipleAgents, unweighted_aggregate_evaluate
 
 __all__ = [
-    'Evaluator',
+    'BaseEvaluator',
     'SotopiaDimensions',
-    'SotopiaDimensions',
-    'EpisodeLLMEvaluator',
     'EvaluationForMultipleAgents',
-    'RuleBasedTerminatedEvaluator',
     'unweighted_aggregate_evaluate',
+    'RuleBasedEvaluator',
+    'LLMEvaluator',
+    'EvaluatorManager',
+    'EvaluatorConfig',
 ]
