+++ tiny_chat/providers/generate.py
@@ -1,6 +1,4 @@
-import os
 from typing import cast
-from urllib.parse import urlparse
 
 from litellm import acompletion, completion
 from litellm.litellm_core_utils.get_supported_openai_params import (
@@ -10,14 +8,6 @@ from litellm.utils import supports_response_schema
 from pydantic import validate_call
 from rich import print
 
-from tiny_chat.generator.output_parsers import (
-    EnvResponse,
-    OutputParser,
-    OutputType,
-    PydanticOutputParser,
-    ScriptOutputParser,
-    StrOutputParser,
-)
 from tiny_chat.messages import (
     ActionType,
     AgentAction,
@@ -42,89 +32,21 @@ from tiny_chat.utils.template import (
     SECOND_PERSON_NARRATIVE_TEMPLATE,
 )
 
+from .output_parsers import (
+    EnvResponse,
+    OutputParser,
+    OutputType,
+    PydanticOutputParser,
+    ScriptOutputParser,
+    StrOutputParser,
+)
+
 
 def _prepare_provider_config(model_name: str) -> tuple[str | None, str | None, str]:
-    """
-    Parse model_name into (api_base, api_key, effective_model_name).
+    """兼容性函数 - 使用providers.utils中的统一实现"""
+    from tiny_chat.providers.utils import prepare_model_config_from_name
 
-    Supported forms:
-    - OpenAI default:                    "gpt-4o-mini"
-    - vLLM (OpenAI-compatible):          "vllm://<model>@<base>" or "vllm://<base>@<model>"
-    - Together (OpenAI-compatible):      "together://<model>"
-    - Custom OpenAI-compatible proxy:    "custom://<model>@<base>" or legacy "custom/<model>@<base>"
-
-    Env vars:
-    - VLLM_API_KEY / TOGETHER_API_KEY / CUSTOM_API_KEY
-    """
-    api_base: str | None = None
-    api_key: str | None = None
-    effective_model = model_name
-
-    def _ensure_v1(url: str) -> str:
-        if not url:
-            return 'http:///v1'
-
-        if not url.startswith(('http://', 'https://')):
-            url = 'http://' + url
-
-        if url.endswith('/v1'):
-            return url
-        trimmed = url[:-1] if url.endswith('/') else url
-        return trimmed + '/v1'
-
-    # vLLM
-    if model_name.startswith('vllm://'):
-        payload = model_name[len('vllm://') :]
-        if '@' not in payload:
-            raise ValueError(
-                "vllm:// requires '<model>@<base_url>' or '<base_url>@<model>'"
-            )
-        a, b = payload.split('@', 1)
-        if urlparse(a).scheme or a.startswith(
-            ('http://', 'https://', 'localhost', '127.0.0.1')
-        ):
-            base, model = a, b
-        else:
-            model, base = a, b
-        api_base = _ensure_v1(base)
-        api_key = os.getenv('VLLM_API_KEY')  # optional
-        effective_model = model
-        return api_base, api_key, effective_model
-
-    # Together
-    if model_name.startswith('together://'):
-        model = model_name[len('together://') :]
-        api_base = 'https://api.together.xyz/v1'
-        api_key = os.getenv('TOGETHER_API_KEY')
-        effective_model = model
-        return api_base, api_key, effective_model
-
-    # Custom proxy (new scheme)
-    if model_name.startswith('custom://'):
-        payload = model_name[len('custom://') :]
-        if '@' not in payload:
-            raise ValueError("custom:// requires '<model>@<base_url>'")
-        model, base = payload.split('@', 1)
-        api_base = base
-        api_key = os.getenv('CUSTOM_API_KEY', 'EMPTY')
-        effective_model = model
-        return api_base, api_key, effective_model
-
-    # Custom proxy (legacy form)
-    if model_name.startswith('custom/'):
-        payload = model_name[len('custom/') :]
-        if '@' not in payload:
-            raise ValueError("custom/ requires '<model>@<base_url>'")
-        model, base = payload.split('@', 1)
-        api_base = base
-        api_key = os.getenv('CUSTOM_API_KEY', 'EMPTY')
-        effective_model = (
-            f'openai/{model}' if not model.startswith('openai/') else model
-        )
-        return api_base, api_key, effective_model
-
-    # default OpenAI
-    return None, None, model_name
+    return prepare_model_config_from_name(model_name)
 
 
 DEFAULT_BAD_OUTPUT_PROCESS_MODEL = 'gpt-4o-mini'
