--- tiny_chat/server/plugins/__init__.py
@@ -1,11 +1,0 @@
-from .base import EvaluatorPlugin
-from .llm_evaluator_plugin import LLMEvaluatorPlugin
-from .manager import PluginManager
-from .rule_based_plugin import RuleBasedPlugin
-
-__all__ = [
-    'EvaluatorPlugin',
-    'PluginManager',
-    'LLMEvaluatorPlugin',
-    'RuleBasedPlugin',
-]
